requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
Markdown==3.2.1
mysqlclient==1.4.6
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
python-decouple==3.3
django-hosts==4.0
djangorestframework==3.11.0
